requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
addict=2.4.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.0.7=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cudatoolkit=11.3.1=ha36c431_9
cycler=0.11.0=pypi_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.40.0=pypi_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
idna=3.4=pypi_0
importlib-metadata=6.6.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.6.1.9=0
libcurand=10.3.2.106=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.3=pypi_0
markdown-it-py=3.0.0=pypi_0
matplotlib=3.7.1=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.6=py38h417a72b_1
mkl_random=1.2.2=py38h417a72b_1
mmcv-full=1.4.0=pypi_0
mmdet=2.20.0=pypi_0
mmengine=0.7.4=pypi_0
model-index=0.1.11=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
opencv-python=4.7.0.72=pypi_0
openh264=2.1.1=h4ff587b_0
openmim=0.3.7=pypi_0
openssl=3.0.8=h7f8727e_0
ordered-set=4.1.0=pypi_0
packaging=23.1=pypi_0
pandas=2.0.2=pypi_0
pillow=9.4.0=py38h6a678d5_0
pip=23.1.2=py38h06a4308_0
pycocotools=2.0.6=pypi_0
pygments=2.15.1=pypi_0
pyparsing=3.0.9=pypi_0
python=3.8.16=h955ad1f_4
python-dateutil=2.8.2=pypi_0
pytorch=1.10.2=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
rich=13.4.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=67.8.0=py38h06a4308_0
shapely=2.0.1=pypi_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
termcolor=2.3.0=pypi_0
terminaltables=3.1.10=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torchaudio=0.10.2=py38_cu113
torchvision=0.11.3=py38_cu113
typing_extensions=4.6.3=py38h06a4308_0
tzdata=2023.3=pypi_0
urllib3=2.0.3=pypi_0
wheel=0.38.4=py38h06a4308_0
xz=5.4.2=h5eee18b_0
yapf=0.33.0=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
